chore: move initial slippage into own hook
